+++ modules/apps/foundation/alloy/alloy-mvc/src/main/java/com/liferay/alloy/mvc/BaseAlloyControllerImpl.java
@@ -14,6 +14,7 @@
 
 package com.liferay.alloy.mvc;
 
+import com.liferay.alloy.mvc.internal.json.web.service.AlloyControllerInvokerManager;
 import com.liferay.alloy.mvc.internal.json.web.service.AlloyMockUtil;
 import com.liferay.counter.kernel.service.CounterLocalServiceUtil;
 import com.liferay.portal.kernel.bean.BeanPropertiesUtil;
@@ -22,6 +23,7 @@ import com.liferay.portal.kernel.dao.search.SearchContainer;
 import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
+import com.liferay.portal.kernel.json.JSONSerializable;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
@@ -1067,6 +1069,18 @@ public abstract class BaseAlloyControllerImpl implements AlloyController {
 		method.invoke(this);
 	}
 
+	protected JSONSerializable invokeAlloyController(
+			String controller, String lifecycle, String action,
+			Object[] parameters)
+		throws Exception {
+
+		AlloyControllerInvokerManager alloyControllerInvokerManager =
+			alloyPortlet.getAlloyInvokerManager();
+
+		return alloyControllerInvokerManager.invokeAlloyController(
+			controller, lifecycle, action, parameters);
+	}
+
 	protected boolean isRespondingTo() {
 		return Validator.isNotNull(format);
 	}
@@ -1275,9 +1289,9 @@ public abstract class BaseAlloyControllerImpl implements AlloyController {
 
 		String simpleClassName = modelClassName.substring(pos + 7);
 
-		String serviceClassName =
-			modelClassName.substring(0, pos) + ".service." + simpleClassName +
-				"LocalService";
+		String serviceClassName = StringBundler.concat(
+			modelClassName.substring(0, pos), ".service.", simpleClassName,
+			"LocalService");
 
 		IdentifiableOSGiService identifiableOSGiService =
 			IdentifiableOSGiServiceUtil.getIdentifiableOSGiService(
@@ -1554,8 +1568,9 @@ public abstract class BaseAlloyControllerImpl implements AlloyController {
 
 		if (log.isDebugEnabled()) {
 			log.debug(
-				"Touch " + portlet.getRootPortletId() + " by including " +
-					touchPath);
+				StringBundler.concat(
+					"Touch ", portlet.getRootPortletId(), " by including ",
+					touchPath));
 		}
 
 		portletContext.setAttribute(