+++ modules/apps/foundation/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/SearchPermissionCheckerImpl.java
@@ -46,6 +46,7 @@ import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.Portal;
+import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.search.configuration.SearchPermissionCheckerConfiguration;
@@ -172,6 +173,16 @@ public class SearchPermissionCheckerImpl implements SearchPermissionChecker {
 				SearchPermissionCheckerConfiguration.class, properties);
 	}
 
+	private void _addGroup(
+		Group group, List<Role> groupRoles,
+		List<UsersGroupIdRoles> usersGroupIdsRoles) {
+
+		if (group != null) {
+			usersGroupIdsRoles.add(
+				new UsersGroupIdRoles(group.getGroupId(), groupRoles));
+		}
+	}
+
 	private void _addPermissionFields(
 			long companyId, long groupId, String className, String classPK,
 			String viewActionId, Document doc)
@@ -238,8 +249,10 @@ public class SearchPermissionCheckerImpl implements SearchPermissionChecker {
 		if (termsCount > permissionTermsLimit) {
 			if (_log.isDebugEnabled()) {
 				_log.debug(
-					"Skipping presearch permission checking due to too many " +
-						"roles: " + termsCount + " > " + permissionTermsLimit);
+					StringBundler.concat(
+						"Skipping presearch permission checking due to too ",
+						"many roles: ", String.valueOf(termsCount), " > ",
+						String.valueOf(permissionTermsLimit)));
 			}
 
 			return null;
@@ -260,9 +273,10 @@ public class SearchPermissionCheckerImpl implements SearchPermissionChecker {
 		if (termsCount > permissionTermsLimit) {
 			if (_log.isDebugEnabled()) {
 				_log.debug(
-					"Skipping presearch permission checking due to too many " +
-						"roles and groups: " + termsCount + " > " +
-							permissionTermsLimit);
+					StringBundler.concat(
+						"Skipping presearch permission checking due to too ",
+						"many roles and groups: ", String.valueOf(termsCount),
+						" > ", String.valueOf(permissionTermsLimit)));
 			}
 
 			return null;
@@ -298,17 +312,20 @@ public class SearchPermissionCheckerImpl implements SearchPermissionChecker {
 				groupRoles.add(siteMemberRole);
 			}
 
-			usersGroupIdsRoles.add(
-				new UsersGroupIdRoles(group.getGroupId(), groupRoles));
+			_addGroup(group, groupRoles, usersGroupIdsRoles);
+
+			_addGroup(group.getStagingGroup(), groupRoles, usersGroupIdsRoles);
 
 			termsCount += groupRoles.size();
 
 			if (termsCount > permissionTermsLimit) {
 				if (_log.isDebugEnabled()) {
 					_log.debug(
-						"Skipping presearch permission checking due to too " +
-							"many roles, groups, and group roles: " +
-								termsCount + " > " + permissionTermsLimit);
+						StringBundler.concat(
+							"Skipping presearch permission checking due to ",
+							"too many roles, groups, and group roles: ",
+							String.valueOf(termsCount), " > ",
+							String.valueOf(permissionTermsLimit)));
 				}
 
 				return null;