+++ modules/sdk/gradle-plugins-source-formatter/CHANGELOG.markdown
@@ -1174,6 +1174,24 @@ another task with the same name is being executed in a parent project.
 - [LPS-75323]: Update the [Liferay Source Formatter] dependency to version
 1.0.484.
 
+## 2.3.14 - 2017-10-25
+
+### Changed
+- [LPS-74849]: Update the [Liferay Source Formatter] dependency to version
+1.0.485.
+
+## 2.3.15 - 2017-10-26
+
+### Changed
+- [LPS-75323]: Update the [Liferay Source Formatter] dependency to version
+1.0.486.
+
+## 2.3.16 - 2017-10-31
+
+### Changed
+- [LPS-75488]: Update the [Liferay Source Formatter] dependency to version
+1.0.487.
+
 [Liferay Source Formatter]: https://github.com/liferay/liferay-portal/tree/master/modules/util/source-formatter
 [LPS-52675]: https://issues.liferay.com/browse/LPS-52675
 [LPS-62970]: https://issues.liferay.com/browse/LPS-62970
@@ -1245,3 +1263,4 @@ another task with the same name is being executed in a parent project.
 [LPS-75254]: https://issues.liferay.com/browse/LPS-75254
 [LPS-75323]: https://issues.liferay.com/browse/LPS-75323
 [LPS-75430]: https://issues.liferay.com/browse/LPS-75430
+[LPS-75488]: https://issues.liferay.com/browse/LPS-75488